third_party/blink/web_tests/http/tests/devtools/compiler-source-mapping-debug-expected.txt

Tests installing compiler source map in scripts panel.


Running: testSetBreakpoint
Script source was shown.
Script execution paused.
Call stack:
    0) handleClick (add-elements.js:15)
    1) clickButton (compiler-source-mapping-debug.js:20)
Script execution resumed.
